+++ includes.h
@@ -4,13 +4,13 @@
 #include <stdbool.h>
 #include <stdarg.h>
 // External symbols from linker
-extern uint32_t __total_pages;
-extern uint32_t _text_size;
-extern uint32_t trampoline_binary;
-extern uint32_t trampoline_size;
+extern uint64_t __total_pages;
+extern uint64_t _text_size;
+extern uint64_t trampoline_binary;
+extern uint64_t trampoline_size;
 
 // External functions
-extern void enable_paging_bootstrap(uint32_t *page_directory);
+extern void enable_paging_bootstrap(uint64_t *page_directory);
 extern void CliHelper();
 extern void StiHelper();
 extern void HltHelper();
@@ -21,7 +21,7 @@ extern uint8_t inb(uint8_t port);
 extern void outb(uint8_t port, uint8_t value);
 extern bool check_interrupts();
 extern void load_tss();
-extern void switch_to_user_mode(uint32_t eip, uint32_t esp);
+extern void switch_to_user_mode(uint64_t eip, uint64_t esp);
 extern uint8_t get_cpuid();
 extern void enable_paging(void* pd);
 extern void enable_sse();
