+++ math.c
@@ -1,16 +1,16 @@
 #include "math.h"
 
-uint32_t pow(uint32_t base, uint32_t exp) {
-    uint32_t result = 1;
-    for (uint32_t i = 0; i < exp; i++) {
+uint64_t pow(uint64_t base, uint64_t exp) {
+    uint64_t result = 1;
+    for (uint64_t i = 0; i < exp; i++) {
         result *= base;
     }
     return result;
 }
 
-uint32_t BiggestBit(uint32_t num) {
-    uint32_t result = 0;
-    for (uint32_t i = 0; i < 32; i++) {
+uint64_t BiggestBit(uint64_t num) {
+    uint64_t result = 0;
+    for (uint64_t i = 0; i < 32; i++) {
         if (num & (1 << i)) {
             result = i;
         }
@@ -18,7 +18,7 @@ uint32_t BiggestBit(uint32_t num) {
     return result;
 }
 
-bool IsPowerOfTwo(uint32_t num) {
+bool IsPowerOfTwo(uint64_t num) {
     if (num == 0) return false;
     return (num & (num - 1)) == 0;
 }
